refactor: 상세페이지를 위한 layout 수정
