+++ src/app/(iTracker)/products/[productId]/page.tsx
@@ -1,0 +1,12 @@
+import { Text } from '@/shared/components/shadcn/Text';
+
+export default function ProductDetail({ params }: { params: { productId: number } }) {
+  console.log(params);
+  return (
+    <main>
+      <Text typography="h4" className="md:text-2xl">
+        상페에지이 입니다.
+      </Text>
+    </main>
+  );
+}
