--- src/app/(iTracker)/products/page.tsx
@@ -1,9 +1,0 @@
-import { Text } from '@/shared/components/shadcn/Text';
-
-export default function Product() {
-  return (
-    <div className="flex-1 w-full flex flex-col gap-10 items-center">
-      <Text typography="h2">Product 페이지 입니다.</Text>
-    </div>
-  );
-}
