+++ src/features/category/components/CategoryList/Item/index.tsx
@@ -12,7 +12,7 @@ export const CategoryItem = ({ categoryItem }: CategoryItemProps) => {
 
   return (
     <li>
-      <Link href={`/products/${categoryItem}`} className="flex flex-col items-center gap-2 cursor-pointer">
+      <Link href={`/category/${categoryItem}`} className="flex flex-col items-center gap-2 cursor-pointer">
         <div className="flex items-center justify-center w-[80px] h-[80px] bg-neutral-100 rounded">
           <Image
             src={`/category/${categoryItem}.png`}
