+++ src/features/product/components/macbook/Item/index.tsx
@@ -16,7 +16,7 @@ export const MacbookProductListItem = ({ productItem }: ProductItemProps) => {
 
   return (
     <li className="w-full">
-      <Link href="" className="flex flex-col gap-2 cursor-pointer">
+      <Link href="/products/1" className="flex flex-col gap-2 cursor-pointer">
         <div className="flex items-center justify-center w-auto h-full rounded-md border-gray-200 border">
           <Image
             src={productItem.imageUrl}
